Upgrade Slate and implement new schema
