+++ lib/core.js
@@ -3,7 +3,7 @@ import Options, { type OptionsFormat } from './options';
 import { isSelectionInBlockquote } from './utils';
 import { wrapInBlockquote, unwrapBlockquote } from './changes';
 
-import { schema, validateNode } from './validation';
+import { schema } from './validation';
 
 /**
  * The core of the plugin, which does not relies on `slate-react`, and includes
@@ -14,7 +14,6 @@ function core(optsParam: OptionsFormat): Object {
 
     return {
         schema: schema(opts),
-        validateNode: validateNode(opts),
 
         utils: {
             isSelectionInBlockquote: isSelectionInBlockquote.bind(null, opts)
