+++ lib/validation/index.js
@@ -1,4 +1,3 @@
 import schema from './schema';
-import validateNode from './validateNode';
 
-export { schema, validateNode };
+export { schema };
