+++ package.json
@@ -6,8 +6,8 @@
   "repository": "git://github.com/GitbookIO/slate-edit-blockquote.git",
   "main": "./dist/index.js",
   "peerDependencies": {
-    "slate": "^0.29.0",
-    "immutable": "^3.8.1"
+    "immutable": "^3.8.1",
+    "slate": "^0.33.0"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
@@ -23,17 +23,17 @@
     "eslint-config-gitbook": "^2.0.3",
     "eslint-plugin-import": "^2.8.0",
     "expect": "^1.20.2",
-    "gh-pages": "^0.11.0",
     "flow-bin": "^0.57.3",
+    "gh-pages": "^0.11.0",
     "http-server": "^0.9.0",
     "immutable": "^3.8.1",
     "mocha": "^3.0.1",
     "react": "^15.3.0",
     "react-dom": "^15.3.0",
     "read-metadata": "^1.0.0",
-    "slate": "^0.29.0",
-    "slate-hyperscript": "^0.4.6",
-    "slate-react": "^0.10.0"
+    "slate": "^0.33.0",
+    "slate-hyperscript": "^0.5.6",
+    "slate-react": "^0.12.0"
   },
   "scripts": {
     "prepublish": "babel ./lib --out-dir ./dist",
@@ -47,5 +47,8 @@
   },
   "keywords": [
     "slate"
-  ]
+  ],
+  "dependencies": {
+    "slate-schema-violations": "^0.1.4"
+  }
 }
