+++ tests/enter-empty-middle/input.yaml
@@ -1,24 +1,24 @@
 document:
   nodes:
-    - kind: block
+    - object: block
       type: blockquote
       nodes:
-        - kind: block
+        - object: block
           type: paragraph
           nodes:
-            - kind: text
+            - object: text
               leaves:
                 - text: P1
-        - kind: block
+        - object: block
           type: paragraph
           key:  '_selection_key'
           nodes:
-            - kind: text
+            - object: text
               leaves:
                 - text: ""
-        - kind: block
+        - object: block
           type: paragraph
           nodes:
-            - kind: text
+            - object: text
               leaves:
                 - text: P3
