+++ tests/schema-items-contain-blocks/input.yaml
@@ -1,17 +1,17 @@
 document:
   nodes:
-    - kind: block
+    - object: block
       type: blockquote
       nodes:
-        - kind: text
+        - object: text
           leaves:
             - text: "1st"
-        - kind: inline
+        - object: inline
           type: link
           nodes:
-            - kind: text
+            - object: text
               leaves:
                 - text: "2nd"
-        - kind: text
+        - object: text
           leaves:
             - text: ""
